pyproject.toml
[project]
name = "cardnews-ai-factory"
version = "0.1.0"
description = "A local base for automating Korean cardnews planning, copywriting, QA, and rendering."
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "anthropic>=0.34.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.0",
  "Jinja2>=3.1.0",
  "playwright>=1.45.0",
  "Pillow>=10.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
